services/ai-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
google-generativeai==0.4.1
upstash-redis==0.15.0
pydantic==2.6.3
python-dotenv==1.0.1
supabase==2.4.0
